# server/docs/swagger.yaml
definitions:
  smserver.addShardRequest:
    properties:
      group:
        description: Group 同一个service需要区分不同种类的shard，这些shard之间不相关的balance到现有container上
        type: string
      manualContainerId:
        type: string
      service:
        description: 为哪个业务app增加shard
        type: string
      shardId:
        type: string
      task:
        description: 业务app自己定义task内容
        type: string
      workerGroup:
        description: WorkerGroup 同一个service需要区分不同种类的container，shard可以指定分配到那一组container上
        type: string
    required:
    - service
    - shardId
    type: object
  smserver.delShardRequest:
    properties:
      service:
        type: string
      shardId:
        type: string
    required:
    - service
    - shardId
    type: object
  smserver.smAppSpec:
    properties:
      createTime:
        type: integer
      maxRecoveryTime:
        description: MaxRecoveryTime 遇到container删除的场景，等待的时间，超时认为该container被清理
        type: integer
      maxShardCount:
        description: MaxShardCount 单container承载的最大分片数量，防止雪崩
        type: integer
      service:
        description: Service 目前app的spec更多承担的是管理职能，shard配置的一个起点，先只配置上service，可以唯一标记一个app
        type: string
    type: object
  smserver.workerRequest:
    properties:
      service:
        description: 为哪个业务app增加worker
        type: string
      worker:
        description: 需要添加的资源，添加后，shard中如果存在WorkerGroup，只会将shard分配到该WorkerGroup下的worker中。
        type: string
      workerGroup:
        description: 在哪个资源组下面添加worker
        type: string
    required:
    - service
    - worker
    - workerGroup
    type: object
info:
  contact: {}
paths:
  /sm/server/add-shard:
    post:
      consumes:
      - application/json
      description: add shard
      parameters:
      - description: param
        in: body
        name: param
        required: true
        schema:
          $ref: '#/definitions/smserver.addShardRequest'
      produces:
      - application/json
      responses:
        "200":
          description: ""
      tags:
      - shard
  /sm/server/add-spec:
    post:
      consumes:
      - application/json
      description: add spec
      parameters:
      - description: param
        in: body
        name: param
        required: true
        schema:
          $ref: '#/definitions/smserver.smAppSpec'
      produces:
      - application/json
      responses:
        "200":
          description: ""
      tags:
      - spec
  /sm/server/add-worker:
    post:
      consumes:
      - application/json
      description: add service worker
      parameters:
      - description: param
        in: body
        name: param
        required: true
        schema:
          $ref: '#/definitions/smserver.workerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: ""
      tags:
      - worker
  /sm/server/del-shard:
    post:
      consumes:
      - application/json
      description: del shard
      parameters:
      - description: param
        in: body
        name: param
        required: true
        schema:
          $ref: '#/definitions/smserver.delShardRequest'
      produces:
      - application/json
      responses:
        "200":
          description: ""
      tags:
      - shard
  /sm/server/del-spec:
    get:
      consumes:
      - application/json
      description: del spec
      parameters:
      - description: param
        in: query
        name: service
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ""
      tags:
      - spec
  /sm/server/del-worker:
    post:
      consumes:
      - application/json
      description: del service worker
      parameters:
      - description: param
        in: body
        name: param
        required: true
        schema:
          $ref: '#/definitions/smserver.workerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: ""
      tags:
      - worker
  /sm/server/get-shard:
    get:
      consumes:
      - application/json
      description: get service all shard
      parameters:
      - description: param
        in: query
        name: service
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ""
      tags:
      - shard
  /sm/server/get-spec:
    get:
      consumes:
      - application/json
      description: get all service
      produces:
      - application/json
      responses:
        "200":
          description: ""
      tags:
      - spec
  /sm/server/get-worker:
    get:
      consumes:
      - application/json
      description: get service workerpool all worker
      parameters:
      - description: param
        in: query
        name: service
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ""
      tags:
      - worker
swagger: "2.0"